+++ team-00/src/main/java/ubc/cosc322/ArrowShot.java
@@ -1,28 +1,28 @@
 package ubc.cosc322;
 
 public class ArrowShot {
-    private int startRow, startCol, endRow, endCol;
+    private byte startRow, startCol, endRow, endCol;
 
-    public ArrowShot(int startRow, int startCol, int endRow, int endCol){
+    public ArrowShot(byte startRow, byte startCol, byte endRow, byte endCol){
         this.startRow = startRow;
         this.startCol = startCol;
         this.endRow = endRow;
         this.endCol = endCol;
     }
 
-    public int getStartRow() {
+    public byte getStartRow() {
         return startRow;
     }
 
-    public int getStartCol() {
+    public byte getStartCol() {
         return startCol;
     }
 
-    public int getEndRow() {
+    public byte getEndRow() {
         return endRow;
     }
 
-    public int getEndCol() {
+    public byte getEndCol() {
         return endCol;
     }
 }
